map is now a fragment
