+++ app/build.gradle
@@ -54,6 +54,8 @@ dependencies {
     // Picasso is a library used to display an image from its URI (cf.: https://github.com/square/picasso)
     implementation 'com.squareup.picasso:picasso:2.71828'
 
+    implementation "androidx.fragment:fragment:1.3.2"
+
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'com.google.android.material:material:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
