+++ app/src/main/res/layout/activity_upload_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:onClick="setDuration"
+    android:id="@+id/activity_upload_activity"
     tools:context=".UploadActivityActivity">
 
     <TextView
